# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Yapay zeka destekli otonom sosyal medya içerik yönetim sistemi API'si.\nBu API, kullanıcıların Instagram hesaplarını bağlamasını, AI ile içerik üretmesini\nve otonom ajanlar aracılığıyla otomatik paylaşım yapmasını sağlar.",
        "title": "Otonom Sosyal Medya Ajan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Destek",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/agent-config": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kullanıcının otonom ajan yapılandırmasını döner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent Config"
                ],
                "summary": "Ajan yapılandırmasını getir",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.AgentConfigResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kullanıcının otonom ajan ayarlarını günceller (AI sağlayıcı, ton, sıklık vb.)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Agent Config"
                ],
                "summary": "Ajan yapılandırmasını güncelle",
                "parameters": [
                    {
                        "description": "Yapılandırma bilgileri",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.UpdateAgentConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Doğrulanmış kullanıcının profil bilgilerini döner",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Kullanıcı profili",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.UserProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/tasks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kullanıcının tüm içerik görevlerini listeler",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Görev listesi",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.TaskListResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "AI tarafından üretilecek yeni bir içerik görevi oluşturur",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Yeni görev oluştur",
                "parameters": [
                    {
                        "description": "Görev bilgileri",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.CreateTaskResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/instagram/callback": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Instagram yetkilendirmesi sonrasında çağrılır. Authorization code'u token'a çevirir,\nInstagram Business hesabını bulur ve Supabase'e kaydeder.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instagram OAuth"
                ],
                "summary": "Instagram OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Instagram authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "CSRF state parametresi",
                        "name": "state",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.InstagramCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Geçersiz code/state veya kullanıcı izin vermedi",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "JWT token geçersiz",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Instagram API hatası",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/instagram/login": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Kullanıcıyı Instagram yetkilendirme sayfasına yönlendirecek URL'i döner.\nDönen authorization_url'e tarayıcıda yönlendirme yapılmalıdır.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Instagram OAuth"
                ],
                "summary": "Instagram OAuth başlat",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.InstagramLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/webhooks/content-scheduled": {
            "post": {
                "description": "İçerik zamanlandığında Supabase tarafından tetiklenir. Worker'a görev gönderir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "İçerik zamanlama webhook'u",
                "parameters": [
                    {
                        "description": "Supabase webhook payload'u",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.WebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/subscription-updated": {
            "post": {
                "description": "Kullanıcı abonelik durumu değiştiğinde Supabase tarafından tetiklenir.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Abonelik güncelleme webhook'u",
                "parameters": [
                    {
                        "description": "Supabase webhook payload'u",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.WebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/user-created": {
            "post": {
                "description": "Supabase'de yeni kullanıcı oluşturulduğunda tetiklenir. Kullanıcı için ilk kurulumu başlatır.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhooks"
                ],
                "summary": "Yeni kullanıcı webhook'u",
                "parameters": [
                    {
                        "description": "Supabase webhook payload'u",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd_api.WebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/cmd_api.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd_api.AgentConfigResponse": {
            "type": "object",
            "properties": {
                "config": {
                    "type": "object",
                    "additionalProperties": true
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "cmd_api.CreateTaskResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "görev oluşturuldu"
                },
                "user_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "cmd_api.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Token geçersiz veya süresi dolmuş"
                },
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
...
        "cmd_api.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "received"
                }
            }
        },
...
        "cmd_api.UpdateAgentConfigRequest": {
            "type": "object",
            "properties": {
                "ai_model": {
                    "type": "string",
                    "example": "gpt-4o"
                },
                "ai_provider": {
                    "type": "string",
                    "enum": [
                        "openai",
                        "gemini"
                    ],
                    "example": "openai"
                },
                "brand_keywords": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "yapay zeka",
                        "teknoloji",
                        "inovasyon"
                    ]
                },
                "is_active": {
                    "type": "boolean",
                    "example": true
                },
                "language": {
                    "type": "string",
                    "example": "tr"
                },
                "posting_frequency": {
                    "type": "string",
                    "enum": [
                        "hourly",
                        "daily",
                        "weekly",
                        "custom"
                    ],
                    "example": "daily"
                },
                "preferred_posting_times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "10:00",
                        "14:00",
                        "19:00"
                    ]
                },
                "target_audience": {
                    "type": "string",
                    "example": "Teknoloji meraklıları, 25-40 yaş"
                },
                "tone": {
                    "type": "string",
                    "example": "professional"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Supabase JWT token. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}